pyproject.toml
[project]
name = "project-0-building-a-specialized-weather-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.6.2",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
